판매자 리스트, 주문 요청 리스트 보여주기 완성

+++ app/src/main/java/ga/washmose/mose/User/UserLaundryItem.java
@@ -15,9 +15,9 @@ public class UserLaundryItem implements Serializable{
     public String location;
     public String summary;
     public ArrayList<ItemData> items;
-    public int latitude, longitude;
+    public double latitude, longitude;
 
-    public UserLaundryItem(String imageUrl, String name, String location, String summary, ArrayList<ItemData> items, int latitude, int longitude) {
+    public UserLaundryItem(String imageUrl, String name, String location, String summary, ArrayList<ItemData> items, double latitude, double longitude) {
         this.imageUrl = imageUrl;
         this.name = name;
         this.location = location;
